+++ libs/btp/tool-header/src/directives/tool-header-group.directive.ts
@@ -1,5 +1,4 @@
-import { BooleanInput, coerceBooleanProperty } from '@angular/cdk/coercion';
-import { Directive, ElementRef, Input, inject } from '@angular/core';
+import { Directive, ElementRef, Input, booleanAttribute, inject } from '@angular/core';
 import { HasElementRef } from '@fundamental-ngx/cdk/utils';
 
 @Directive({
@@ -13,12 +12,12 @@ import { HasElementRef } from '@fundamental-ngx/cdk/utils';
 })
 export class ToolHeaderGroupDirective implements HasElementRef {
     /** @hidden */
-    @Input({ transform: coerceBooleanProperty })
-    center: BooleanInput;
+    @Input({ transform: booleanAttribute })
+    center = false;
 
     /** @hidden */
-    @Input({ transform: coerceBooleanProperty })
-    actions: BooleanInput;
+    @Input({ transform: booleanAttribute })
+    actions = false;
 
     /** @hidden */
     readonly elementRef = inject<ElementRef<HTMLElement>>(ElementRef);