+++ libs/platform/form/multi-input/multi-input.component.ts
@@ -263,8 +263,6 @@ export class PlatformMultiInputComponent extends BaseMultiInput implements OnIni
         if (this.mobile) {
             this._setUpMobileMode();
         }
-
-        this.tokenizer._showOverflowPopover = false;
     }
 
     /** @hidden Method to emit change event */
@@ -307,6 +305,10 @@ export class PlatformMultiInputComponent extends BaseMultiInput implements OnIni
 
     /** @hidden */
     addOnButtonClick(event: Event): void {
+        if (isFunction(this.addOnButtonClickFn)) {
+            this.addOnButtonClickFn();
+            return;
+        }
         if (!this.openDropdownOnAddOnClicked && this.isOpen) {
             this.showList(false);
         } else if (this.openDropdownOnAddOnClicked) {
@@ -315,11 +317,6 @@ export class PlatformMultiInputComponent extends BaseMultiInput implements OnIni
         }
 
         this.addOnButtonClicked.emit(event);
-
-        if (isFunction(this.addOnButtonClickFn)) {
-            this.addOnButtonClickFn();
-            return;
-        }
     }
 
     /** @hidden */